capture: generate tcpdump command-line as fallback
previously there were two ways to capture traffic with hotwire:
1. on linux with pkexec, with the capture button
2. opening a fifo and manually calling tcpdump on it

The second option was needed for flatpak and OSX. Now in the second case
we still offer the capture button, but it'll generate the tcpdump
command to run for you. There is also a preference option to use this
way on linux native if you have to for some reason.